pyproject.toml
[tool.poetry]
name = "trucktrack"
version = "0.1.0"
description = ""
authors = ["Sherif Ragimov <ragimov700@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.11"
flake8 = "^7.0.0"
isort = "^5.13.2"
djangorestframework = "^3.15.1"
geopy = "^2.4.1"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
